@@ -1,13 +1,9 @@
 package org.reactome.server.analysis.service.controller;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.reactome.server.analysis.AppTests;
+import org.junit.jupiter.api.Test;
+import org.reactome.server.analysis.service.AppTests;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.request.MockMultipartHttpServletRequestBuilder;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
@@ -19,12 +15,10 @@ import java.util.Map;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration({"file:src/test/resources/mvc-dispatcher-servlet-test.xml"})
-@WebAppConfiguration
+
 public class IdentifiersControllerTest extends AppTests {
 
-    private String overrepresentationFileContent = "#GBM Uniprot\n" +
+    private final String overrepresentationFileContent = "#GBM Uniprot\n" +
             "P01023\n" +
             "Q99758\n" +
             "O15439\n" +
@@ -40,7 +34,7 @@ public class IdentifiersControllerTest extends AppTests {
             "Q96GD4\n" +
             "Q13145";
 
-    private String expressionFileContent = "#Gene symbol\t#Somatic\t#Germline\tTotal\n" +
+    private final String expressionFileContent = "#Gene symbol\t#Somatic\t#Germline\tTotal\n" +
             "TP53        \t       8\t        5\t   18\n" +
             "APC         \t       6\t        6\t   17\n" +
             "TERT        \t       9\t        1\t   15\n" +
@@ -52,10 +46,9 @@ public class IdentifiersControllerTest extends AppTests {
             "CTNNB1      \t       5\t        0\t   10\n" +
             "ERBB2       \t       5\t        0\t   10";
 
-    private MockMultipartFile overrepresentationFile = new MockMultipartFile("file", "tuple-mentha-psimitab-ex.txt", "multipart/form-data", overrepresentationFileContent.getBytes());
-    private MockMultipartFile expressionFile = new MockMultipartFile("file", "tuple-mentha-psimitab-ex.txt", "multipart/form-data", expressionFileContent.getBytes());
-
-    private Map<String, Object> params = new HashMap<>();
+    private final MockMultipartFile overrepresentationFile = new MockMultipartFile("file", "tuple-mentha-psimitab-ex.txt", "multipart/form-data", overrepresentationFileContent.getBytes());
+    private final MockMultipartFile expressionFile = new MockMultipartFile("file", "tuple-mentha-psimitab-ex.txt", "multipart/form-data", expressionFileContent.getBytes());
+    private final Map<String, Object> params = new HashMap<>();
 
     // this is a constructor
     public IdentifiersControllerTest() {
@@ -94,12 +87,11 @@ public class IdentifiersControllerTest extends AppTests {
         files.add(overrepresentationFile);
         files.add(expressionFile);
 
-        MockMultipartHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.fileUpload("/identifiers/form/projection");
+        MockMultipartHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.multipart("/identifiers/form/projection");
 
         for (MockMultipartFile file : files) {
             requestBuilder.file(file);
         }
-
         this.getMockMvc().perform(requestBuilder
                 .contentType(MediaType.MULTIPART_FORM_DATA)
                 .param("interactors", "false")
@@ -121,12 +113,11 @@ public class IdentifiersControllerTest extends AppTests {
         files.add(overrepresentationFile);
         files.add(expressionFile);
 
-        MockMultipartHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.fileUpload("/identifiers/form/");
+        MockMultipartHttpServletRequestBuilder requestBuilder = MockMvcRequestBuilders.multipart("/identifiers/form/");
 
         for (MockMultipartFile file : files) {
             requestBuilder.file(file);
         }
-
         this.getMockMvc().perform(requestBuilder
                 .contentType(MediaType.MULTIPART_FORM_DATA)
                 .param("interactors", "false")
@@ -139,14 +130,12 @@ public class IdentifiersControllerTest extends AppTests {
                 .param("includeDisease", "true"))
                 .andExpect(status().isOk())
                 .andReturn();
-
     }
 
     @Test
     public void getPostURLToHuman() throws Exception {
         String uniProtACFileUrl = "https://raw.githubusercontent.com/Chuqiaoo/reactome-analysis-service-testing-files/master/uniprotACs.txt";
         mockMvcPostResult("/identifiers/url", uniProtACFileUrl, params);
-
     }
 
     @Test
