@@ -1,24 +1,19 @@
 package org.reactome.server.analysis.service.controller;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.reactome.server.analysis.AppTests;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.reactome.server.analysis.service.AppTests;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
 
 import java.util.HashMap;
 import java.util.Map;
 
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration({"file:src/test/resources/mvc-dispatcher-servlet-test.xml"})
-@WebAppConfiguration
+
 public class TokenControllerTest extends AppTests {
 
-    @Before
-    public void prepare(){
+    @BeforeEach
+    public void prepare() {
         generateToken("P02452 P08123 P02461 P12110 P49674 P35222 P09668 Q9NQC7");
     }
 
@@ -34,7 +29,7 @@ public class TokenControllerTest extends AppTests {
         params.put("order", "ASC");
         params.put("pValue", 1);
         params.put("includeDisease", false);
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, params);
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", params);
     }
 
     @Test
@@ -47,7 +42,6 @@ public class TokenControllerTest extends AppTests {
         params.put("includeDisease", false);
         String input = AppTests.stId;
         mockMvcPostResult(url, input, params);
-
     }
 
     @Test
@@ -57,7 +51,7 @@ public class TokenControllerTest extends AppTests {
         params.put("sortBy", "ENTITIES_PVALUE");
         params.put("order", "ASC");
         params.put("resource", "TOTAL");
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, params);
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", params);
     }
 
     @Test
@@ -70,14 +64,13 @@ public class TokenControllerTest extends AppTests {
         params.put("resource", "TOTAL");
         params.put("pValue", 1);
         params.put("includeDisease", true);
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, params);
-
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", params);
     }
 
     @Test
     public void getTokenHitEntitiesPathway() throws Exception {
         String url = String.format("/token/%s/found/all/%s", AppTests.token, AppTests.stId);
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, "resource", "TOTAL");
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", "resource", "TOTAL");
     }
 
     @Test
@@ -90,14 +83,14 @@ public class TokenControllerTest extends AppTests {
     @Test
     public void getTokenIdentifiersPathway() throws Exception {
         String url = String.format("/token/%s/found/entities/%s", AppTests.token, AppTests.stId);
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, "page", "1");
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", "page", "1");
     }
 
     /* API ignored*/
     @Test
     public void getTokenSummaryPathway() throws Exception {
         String url = String.format("/token/%s/summary/%s", AppTests.token, AppTests.stId);
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, "page", "1");
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", "page", "1");
     }
 
     @Test
@@ -107,19 +100,19 @@ public class TokenControllerTest extends AppTests {
         params.put("pageSize", 20);
         params.put("page", 1);
         params.put("resource", "TOTAL");
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, params);
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", params);
     }
 
     @Test
     public void getNotFoundIdentifiers() throws Exception {
         String url = String.format("/token/%s/notFound", AppTests.token);
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, null);
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", null);
     }
 
     @Test
     public void getTokenFilterPathwayReactions() throws Exception {
         String url = String.format("/token/%s/reactions/%s", AppTests.token, AppTests.stId);
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, "resource", "TOTAL");
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", "resource", "TOTAL");
     }
 
     @Test
@@ -133,13 +126,12 @@ public class TokenControllerTest extends AppTests {
         params.put("species", 48887);
         params.put("includeDisease", false);
         mockMvcPostResult(url, input, params);
-
     }
 
     @Test
     public void getResources() throws Exception {
         String url = String.format("/token/%s/resources", AppTests.token);
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, null);
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", null);
     }
 
     @Test
@@ -151,6 +143,6 @@ public class TokenControllerTest extends AppTests {
         params.put("pValue", 1);
         params.put("species", 48887);
         params.put("includeDisease", true);
-        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE, params);
+        mockMvcGetResult(url, MediaType.APPLICATION_JSON_VALUE + ";charset=UTF-8", params);
     }
 }
