@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.servlet.http.HttpServletRequest;
 
 @Controller
-@Api(tags = "species", description= "Species comparison", position = 4)
+@Api(tags = "species")
 @RequestMapping(value = "/species")
 public class SpeciesController {
 
@@ -46,7 +46,7 @@ public class SpeciesController {
                                                 HttpServletRequest request) {
         Long from = SpeciesNodeFactory.getHumanNode().getSpeciesID(); //For the time being let's do only human ;)
         return controller.compareSpecies(from, species, request)
-                .filterPathways(resource, pValue, false, min, max)
+                .filterPathways(resource, pValue, true, min, max)
                 .getResultSummary(sortBy, order, resource, pageSize, page);
     }
 
